# target/classes/openapi.yaml
openapi: 3.0.0
info:
  title: Wegfest API
  description: API zur Kommunikation mit der Oculus Quest.
  version: 0.0.1
servers:
  - url: http://localhost:8080/api
    description: URL für die lokale Testumgebung
components:
  schemas:
...
    Answer:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
    Question:
      type: object
      properties:
        _id:
          type: string
        question:
          type: string
        answers:
          type: array
          items: 
            $ref: '#/components/schemas/Answer'
    Scene:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        description: 
          type: string
        difficulty:
          type: string
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
...
paths:
...
  /scenes:
    get:
      tags:
        - Scenes
      summary: Liefert alle Szenen zurück
      description: Liefert alle in der Datenbank vorhandenen Szenen zurück
      operationId: getAllScenes
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Scene'
        '401':
          description: Authorization Information is Missing or Invalid
                
    post:
      tags:
        - Scenes
      summary: Erstellt eine neue Szene
      description: Persistiert die Szene mit den übergebenen Attributen in der Datenbank
      operationId: createScene
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Scene'
      responses:
        '201':
          description: Scene Created Successfully
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: Authorization Information is Missing or Invalid
  /scenes/{id}:
    get:
      tags:
        - Scenes
      summary: Liefert die Szene mit der angegebenen ID zurück
      description: Liefert die Szene mit der angegebenen ID zurück
      operationId: getScene
      parameters:
        - in: path
          name: id
          description: Scene ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '401':
          description: Authorization Information is Missing or Invalid
        '404':
          description: No Scene With Corresponding ID Found
  
    put:
      tags:
        - Scenes
      summary: Aktualisiert die Szene mit der angegbenen ID
      description: Aktualisiert die Szene mit den entsprechenden Werten aus dem Request Body
      operationId: updateScene
      parameters:
        - in: path
          name: id
          description: Scene ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Scene'
      responses:
        '201':
          description: Scene Created Successfully
        '204':
          description: Scene Updated Successfully
        '401':
          description: Authorization Information is Missing or Invalid
        '404':
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string
                
    delete:
      tags:
        - Scenes
      summary: Löscht die Szene mit der angegebenen ID
      description: Löscht die Szene mit der angegebenen ID aus der Datenbank
      operationId: deleteScene
      parameters:
        - in: path
          name: id
          description: Scene ID
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Scene Deleted Successfully
        '401':
          description: Authorization Information is Missing or Invalid
        '404':
          description: No Scene With Corresponding ID Found
  /questions:
    get:
      tags:
        - Questions
      summary: Liefert alle Fragen zurück
      description: Liefert alle in der Datenbank vorhandenen Fragen zurück
      operationId: getAllQuestions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
        '401':
          description: Authorization Information is Missing or Invalid
    
    post:
      tags:
        - Questions
      summary: Erstellt eine neue Frage
      description: Persistiert die Frage mit den übergebenen Attributen in der Datenbank
      operationId: createQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Question'
      responses:
        '201':
          description: Question Created Successfully
        '401':
          description: Authorization Information is Missing or Invalid
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string
  /questions/{id}:
    get:
      tags:
        - Questions
      summary: Liefert die Frage mit der angegebenen ID zurück
      description: Liefert die Frage mit der angegebenen ID zurück
      operationId: getQuestion
      parameters:
        - in: path
          name: id
          description: Question ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '401':
          description: Authorization Information is Missing or Invalid
        '404':
          description: No Question With Corresponding ID Found
    
    put:
      tags:
        - Questions
      summary: Aktualisiert die Frage mit der angegebenen ID
      description: Aktualisiert die Frage mit den entsprechenden Werten aus dem Request Body
      operationId: updateQuestion
      parameters:
        - in: path
          name: id
          description: Question ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Question'
      responses:
        '201':
          description: Question Created Successfully
        '204':
          description: Question Updated Successfully
        '401':
          description: Authorization Information is Missing or Invalid
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                type: string
            
    delete:
      tags:
        - Questions
      summary: Löscht die Frage mit der angegebenen ID
      description: Löscht die Frage mit der angegebenen ID aus der Datenbank
      operationId: deleteQuestion
      parameters:
        - in: path
          name: id
          description: Question ID
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Question Deleted Successfully
        '401':
          description: Authorization Information is Missing or Invalid
        '404':
          description: No Question With Corresponding ID Found